fix: correct rule's kind & stmt
Created rule with `NewRule()` is different
to the rule created with `ruleFromExpr()`
(when parsing input data).

In the first case (`NewRule()`) kind is set to
`build.Ident{Name: string}` while it's parsed
expression in the second case (`ruleFromExpr()`).

For rules packed in a struct like
`selects.config_setting_group` there is no way
to create/generate the same rule as the one
parsed from the file as both will have
different kinds.

This change unifies kinds for two ways of
creating rules and correctly sets the identifier
while fixing loads (`FixLoads()`).